fixed search and other things (see description)
Listing:
initialized the questions array in Listing.

ListingService:
 added a line in createListing function in ListingService to add the listing in the instance, changed the searchItems func to searchListings and made the required changes.

Main:
initialized the LoginUserEmail to null.

The Listing data needed initialization before calling thr report and the search function as it was only called IN these functions, so i had to make a dummy ListingData obj in the start of the main.

made some changes in the search function according to the changes in searchListings, Added a condition to check if there is a logged in user to report or not in the report function, made it so input takes whole lines (lina helped me on this) and ignored case in letters.